Fix glitch on class creation - fixes #3
Quit automatically navigating to the newly created class.

I don't know why this was causing issues, but a one-line fix that makes
a rare event slightly more painful is better than debugging Meteor
itself.  Maybe Hamwerk was trying to load the assignments in the new
class before the class was created?