Always pull all assignments - fix #12 and fix #19
Previously, we were calling `Meteor.subscribe` to limit which
assignments we were downloading.  This would be fine if Offline.save()
could still get the full list, but this was limiting the cached stuff to
whatever assignment was most recently viewed in some circumstances.
Since we fixed that, we can just run `Offline.save()` every second.